fix all warnings

+++ src/os345lc3.c
@@ -44,7 +44,7 @@ int getCharacter(void);
 
 extern unsigned short int *getMemAdr(int va, int rwFlg);
 extern void setFrameTableBits(int flg, int sf, int ef);
-extern int accessPage(int pnum, int frame, int rwnFlg);
+extern long accessPage(int pnum, int frame, int rwnFlg);
 
 // ***********************************************************************
 // ***********************************************************************
